move brazilian patterns to shared
