+++ client/src/components/results-section.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import { Progress } from "@/components/ui/progress";
-import { getRiskLevelColor, getRiskLevelBadgeColor, getRiskLevelText } from "@/lib/brazilian-patterns";
+import { getRiskLevelColor, getRiskLevelBadgeColor, getRiskLevelText } from "@shared/brazilian-patterns";
 import { apiRequest } from "@/lib/queryClient";
 import jsPDF from 'jspdf';
 import autoTable from 'jspdf-autotable';
